requirements.txt
<<<<<<< HEAD
streamlit==1.54.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
pandas==2.1.4
=======
streamlit==1.54.0
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
pandas==2.1.4
>>>>>>> d7aab021af56edae4ca8a7650116415220551368
